# Replace TCP connection buffering with backpressure